requirements.txt
Flask==0.10.1
Flask-Admin==1.0.8
Flask-Assets==0.8
Flask-Cache==0.12
Flask-Compress==0.10.0
Flask-Login==0.2.7
Flask-RESTful==0.2.5
Flask-WTF==0.10.3
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==2.8.1
SQLAlchemy==0.9.2
WTForms==2.0.1
Werkzeug==0.9.6
argparse==1.2.1
certifi==2015.04.28
ecdsa==0.13
itsdangerous==0.24
maxcdn==0.0.7
newrelic==2.0.0.1
oauthlib==0.7.2
paramiko==1.15.2
ply==3.4
pyOpenSSL==0.13.1
pycrypto==2.6.1
pygeoip==0.2.7
pymongo==2.6.2
python-dateutil==2.1
pytz==2013.7
redis==2.9.1
requests==2.3.0
requests-oauthlib==0.5.0
six==1.4.1
slimit==0.8.1
tldextract==1.2.2
webassets==0.8
wsgiref==0.1.2
xmltodict==0.7.0
